requirements.txt
telethon>=1.36
pyyaml>=6.0
